♻️ refactor: extracts shared image and overview utilities
